--- a/ebos/eclcpgridvanguard.hh
+++ b/ebos/eclcpgridvanguard.hh
@@ -193,10 +193,6 @@ public:
             {
                 const auto wells = this->schedule().getWellsatEnd();
 
-
-                auto& eclState = static_cast<ParallelEclipseState&>(this->eclState());
-                const EclipseGrid* eclGrid = nullptr;
-
                 try
                 {
                     auto& eclState = dynamic_cast<ParallelEclipseState&>(this->eclState());
@@ -209,8 +205,7 @@ public:
 
                     PropsCentroidsDataHandle<Dune::CpGrid> handle(*grid_, eclState, eclGrid, this->centroids_,
                                                                   cartesianIndexMapper());
-                    defunctWellNames_ = std::get<1>(grid_->loadBalance(handle, edgeWeightsMethod, &wells, faceTrans.data()));
-                    //defunctWellNames_ = std::get<1>(grid_->loadBalance(edgeWeightsMethod, ownersFirst, &wells, faceTrans.data()));
+                    defunctWellNames_ = std::get<1>(grid_->loadBalance(handle, edgeWeightsMethod, &wells, faceTrans.data(), ownersFirst));
                 }
                 catch(const std::bad_cast& e)
                 {