This PR require https://github.com/OPM/opm-grid/pull/430 in opm-grid, and in addition to what is contained there, the following is changed:

- Command line argument --owners-first, that turn on/off "owners first" ordering of cells in the parallel grid. Default is off.

- SpMV, inner product and ILU that utilise the "owners first" ordering, and therefore avoids some superfluous computations.

On dual Intel Xeon E5-2670 2.6GHz 8-core CPU nodes this time, some modest improvement in parallel performance was achieved on Norne:

| P | master: time / iter     | owners-first: time / iter  |
|---|--------------------| -----------------------|
| 2  |  426.57 / 25339     |    408.73  /  25179        |
| 4  |     259.07 / 25571    |      245.40 / 25209      |
| 8  |     181.15 / 25026     |     174.39 / 25656       |
| 16 |   129.76 / 25738     |     119.90 / 25227         |
| 32 |    98.59 / 26317     |     87.27 / 26113            |
| 64 |    77.36 / 26769     |     69.73 / 26726          |

The change in iteration count is due to the reordering and ILU being sensitive to that.

This PR was created for testing on models I don't have access to. A more elegant implementation that avoids having to copy the matrix, by altering the assembly part of the code would probably achieve similar performance without having to deal with the "owners first" ordering stuff.