@blattms 2019-12-11
jenkins run norne_parallel opm-grid=430 please