@andrthu 2020-03-11
SpMV and ILU can utilise the owners first ordering, while CPR (or AMG) can not. Therefore we need to modify the matrix by setting copy rows to 1 on the diagonal when using CPR and AMG, even if --owner-cells-first=true.