@blattms 2020-03-13
Would it make sense to add a warning if --owner-cellls-first=true is ignored? Currently one cannot judge this without inside knowledge.