@blattms 2020-03-13
jenkins run norne_parallel opm-grid=430 please