@blattms 2020-03-13
Shoot, did not run serial norne.