@blattms 2020-03-13
But everything looks reasonable with the default (--owner-cells-first=true) which should not change anything. Merging with upstream.